Hak Akses, Sign In, Login
17 Juni 2023

+++ app/Http/Controllers/admin/PenggunaController.php
@@ -9,9 +9,14 @@ use Illuminate\Support\Facades\DB;
 
 class PenggunaController extends Controller
 {
-  
+
     public function index()
     {
+        // Cek hak akses jika dia bukan admin, maka akan diarahkan ke Landingpage
+        if (auth()->user()->id_level != 1) {
+            return redirect('/')->withErrors('Anda tidak memiliki hak akses.');
+        }
+
         $user = User::all();
 
         return view("admin.pengguna.index", compact('user'), [
@@ -19,10 +24,14 @@ class PenggunaController extends Controller
         ]);
     }
 
-      public function hapus_data_pengguna($id)
+    public function hapus_data_pengguna($id)
     {
+        // Cek hak akses jika dia bukan admin, maka akan diarahkan ke Landingpage
+        if (auth()->user()->id_level != 1) {
+            return redirect('/')->withErrors('Anda tidak memiliki hak akses.');
+        }
+
         DB::table('User')->where('id', $id)->delete();
         return redirect('Data Pengguna');
     }
-  
 }
